fix issue with youtube playlits not loading the right number of videos fix tv home to refresh on server change
